+++ packages/cubejs-athena-driver/src/AthenaDriver.ts
@@ -8,7 +8,7 @@ import {
   DriverInterface,
   QueryOptions, StreamOptions,
   StreamTableData, TableName
-} from '@cubejs-backend/query-orchestrator';
+} from '@cubejs-backend/base-driver';
 import { checkNonNullable, getEnv, pausePromise, Required } from '@cubejs-backend/shared';
 import * as SqlString from 'sqlstring';
 import { AthenaClientConfig } from '@aws-sdk/client-athena/dist-types/AthenaClient';