+++ packages/cubejs-ksql-driver/src/KsqlDriver.ts
@@ -2,7 +2,7 @@
 import {
   BaseDriver,
   DriverInterface,
-} from '@cubejs-backend/query-orchestrator';
+} from '@cubejs-backend/base-driver';
 import { format as formatSql } from 'sqlstring';
 import axios, { AxiosResponse } from 'axios';
 import { Mutex } from 'async-mutex';