+++ packages/cubejs-prestodb-driver/driver/PrestoDriver.js
@@ -2,7 +2,7 @@ const presto = require('presto-client');
 const {
   map, zipObj, prop, concat
 } = require('ramda');
-const { BaseDriver } = require('@cubejs-backend/query-orchestrator');
+const { BaseDriver } = require('@cubejs-backend/base-driver');
 const SqlString = require('sqlstring');
 
 class PrestoDriver extends BaseDriver {
@@ -13,7 +13,7 @@ class PrestoDriver extends BaseDriver {
     return 2;
   }
 
-  constructor(config) {
+  constructor(config = {}) {
     super();
 
     this.config = {