+++ packages/cubejs-query-orchestrator/src/index.ts
@@ -1,2 +1,4 @@
-export * from './driver';
 export * from './orchestrator';
+
+// reexport for backward compatibility
+export * from '@cubejs-backend/base-driver';