+++ packages/cubejs-query-orchestrator/src/orchestrator/PreAggregations.ts
@@ -14,22 +14,18 @@ import {
   utcToLocalTimeZone,
 } from '@cubejs-backend/shared';
 
-import { cancelCombinator, SaveCancelFn } from '../driver/utils';
+import { cancelCombinator, SaveCancelFn, DriverInterface, BaseDriver,
+  DownloadTableData,
+  InlineTable,
+  StreamOptions,
+  UnloadOptions } from '@cubejs-backend/base-driver';
 import { RedisCacheDriver } from './RedisCacheDriver';
 import { LocalCacheDriver } from './LocalCacheDriver';
 import { Query, QueryCache, QueryTuple, QueryWithParams } from './QueryCache';
 import { ContinueWaitError } from './ContinueWaitError';
 import { DriverFactory, DriverFactoryByDataSource } from './DriverFactory';
 import { CacheDriverInterface } from './cache-driver.interface';
-import {
-  BaseDriver,
-  DownloadTableData,
-  InlineTable,
-  StreamOptions,
-  UnloadOptions
-} from '../driver';
 import { QueryQueue } from './QueryQueue';
-import { DriverInterface } from '../driver/driver.interface';
 import { LargeStreamWarning } from './StreamObjectsCounter';
 
 /// Name of the inline table containing the lambda rows.
@@ -1574,7 +1570,7 @@ export class PreAggregations {
 
   private cacheDriver: CacheDriverInterface;
 
-  public externalDriverFactory: any;
+  public externalDriverFactory: DriverFactory;
 
   public structureVersionPersistTime: any;