+++ packages/cubejs-testing/test/driver-firebolt.test.ts
@@ -1,3 +1,14 @@
-import { executeTestSuiteFor } from './driver-test-suite';
+import { mainTestSet } from './driverTests/testSets';
+import { executeTestSuite } from './driver-test-suite';
 
-executeTestSuiteFor('firebolt');
+executeTestSuite({
+  type: 'firebolt',
+  tests: mainTestSet,
+  
+});
+
+executeTestSuite({
+  type: 'firebolt',
+  tests: mainTestSet,
+  config: { CUBEJS_EXTERNAL_DEFAULT: 'true' }
+});